Fix HrrSign definition for absorbing elements
The sign definition was not working for absorbing elements (e.g.
`np.ones(d)`) with even dimensionalities. The resulting sign had a
Nyquist coefficient of zero which was not allowed to be converted back
into a sign vector. This is resolved by using the dc_sign as
nyquist_sign if the nyquist_sign happens to be zero.

This also simplifies the implementation an bit because certain error
cases (e.g. odd dimensionality with non-zero dc_sign, but zero
nyquist_sign) cannot happen anymore. (Well, except someone manipulates
the instance after instantiation, but one shouldn't do that.)