Fix HRR and default abs definition
While multiplying with the sign works for real numbers, we actually
need to multiply/bind with the inverse of the sign if we look at the
analogy of complex numbers where `abs(z) = (1 / sign(z)) * z`. It also
makes sense, if you think about it because for the `abs` function we
want to remove the sign, thus multiplying with the inverse. It does not
hold in general that a sign is its own inverse (as for real numbers
where 1 / -1 = -1).

This also eliminates the weird `roll` in the HrrAlgebra test for abs
and producing the expected relationship `sign(v) * abs(v) = v`.